# backend/config/swagger.json
{
    "swagger": "2.0",
...
    "info": {
        "description": "API for calculating your Carbon footprint",
        "version": "0.1.0",
        "title": "Footprint API",
        "termsOfService": "https://footprint.codefor.nl/terms/",
        "contact": {
            "email": "apiteam@codefor.nl"
        },
        "license": {
            "name": "GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        }
    },
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/person": {
            "post": {
                "tags": [
                    "person"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PostPerson"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "footprint_auth": [
                            "write:travel",
                            "read:travel"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "person"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "footprint_auth": [
                            "write:person",
                            "read:person"
                        ]
                    }
                ]
            }
        },
        "/person/{id}": {
            "get": {
                "tags": [
                    "person"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "400": {
                        "description": "Invalid ID"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "person"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "footprint_auth": [
                            "write:person",
                            "read:person"
                        ]
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "footprint_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
...
        "Person": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "PostPerson": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}